VZ-5858 minimal metrics template attempt
